test(cli): extract smoke-test boilerplate to tests/support/mod.rs
Five smoke tests (`cli_flow`, `lint_smoke`, `clean_smoke`,
`rename_smoke`, `undo_redo_smoke`) each carried their own
`binary_path` / `init_project` / `touch` / `write_file` / `run`
helpers. The bodies were byte-for-byte identical except for the
project name passed to `progest init`.

Move them to `crates/progest-cli/tests/support/mod.rs` (Cargo's
standard pattern for shared integration-test code). Each smoke file
now declares `mod support;` and re-binds `init_project` locally to
the test-specific project name so the per-test signature is
unchanged. `support::run` returns `Output` so existing callers that
inspect stdout/stderr keep working.

Net: -100 +86 lines (one new module, five duplicate helper sets
deleted).
